feat(cpu): qwen CPU 三缺陷修复+全引擎成绩单(10 组零报错)
- DenseVQEmbedding uint16 行索引转 int64(torch 高级索引限制)
- fla chunk 实例绑定无设备检查→实例级设备分派(CPU 张量不再直入
  Triton);fla FusedRMSNormGated(Triton-only)CPU 换 torch 等价实现
- bench_any_model 加 --cpu 与 prefill/decode 分离计量
- 成绩单(docs/引擎成绩单-20260820.md):
  CUDA: qwen MTP 66.51/plain 46.15@15.9GiB,DSV4 30.02,GLM int4
  8.12(prefill 106)/fp8 5.28,Kimi 3.42
  CPU: qwen 7.06(prefill 17.5),DSV4 8.70,GLM 0.46,Kimi 1.34
- 回归:qwen CUDA 三件套全绿(MTP 56.98 方差带/acceptance 5/12+2/12
  基线/plain 46.15 finite);10 组运行零报错

+++ engine/CCCP-Engine/cccp/dense_vq.py
@@ -1649,7 +1649,12 @@ class DenseVQEmbedding(nn.Module):
                 blocks=self.blocks,
                 bits=self.bits,
             )
-            values = self.codebook[indices].reshape(-1, self.embedding_dim)
+            # 16bit 载荷的行索引是 uint16——torch 高级索引只收
+            # long/int/byte/bool,CPU 嵌入查表需转 int64(CUDA 路径
+            # 走 fused 算子不经此处;CPU 模式首跑实证)。
+            values = self.codebook[indices.long()].reshape(
+                -1, self.embedding_dim
+            )
             values = values.to(torch.bfloat16)
         return values.reshape(*shape, self.embedding_dim)
